requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiolimiter==1.1.0
aiosignal==1.3.1
aliyunsdkcore==1.0.3
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
anytree==2.12.1
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
autograd==1.7.0
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
azure-search-documents==11.5.2
azure-storage-blob==12.24.0
backoff==2.2.1
bcrypt==4.2.0
beartype==0.18.5
botorch==0.12.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.18
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
ConfigSpace==1.2.0
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
Cython==3.0.11
dask==2024.11.2
dask-jobqueue==0.9.0
dataclasses-json==0.6.7
datashaper==0.0.49
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
deprecation==2.1.0
devtools==0.12.2
diskcache==5.6.3
distributed==2024.11.2
distro==1.9.0
durationpy==0.9
dynaconf==3.2.6
emcee==3.1.6
environs==11.2.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.5
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gensim==4.3.3
google-auth==2.36.0
googleapis-common-protos==1.66.0
gpytorch==1.13
graphrag==0.4.1
graspologic==3.4.1
graspologic-native==1.2.1
greenlet==3.1.1
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
hyppo==0.4.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipython==8.29.0
isodate==0.7.2
jaxtyping==0.2.19
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.1
jmespath==0.10.0
joblib==1.4.2
json_repair==0.30.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kubernetes==31.0.0
lancedb==0.13.0
langchain==0.2.6
langchain-anthropic==0.3.0
langchain-chroma==0.1.2
langchain-community==0.2.6
langchain-core==0.2.10
langchain-openai==0.1.11
langchain-text-splitters==0.2.2
langgraph==0.1.5
langgraph-sdk==0.1.36
langsmith==0.1.143
lark==1.2.2
lightgbm==4.5.0
linear-operator==0.5.3
llvmlite==0.43.0
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
msal==1.31.0
msal-extensions==1.2.0
msgpack==1.1.0
multidict==6.1.0
multipledispatch==1.0.0
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
onnxruntime==1.20.0
openai==1.54.4
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-instrumentation==0.49b1
opentelemetry-instrumentation-asgi==0.49b1
opentelemetry-instrumentation-fastapi==0.49b1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
opentelemetry-util-http==0.49b1
opt_einsum==3.4.0
orjson==3.10.11
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
partd==1.4.2
patsy==1.0.1
pexpect==4.9.0
pillow==11.0.0
Platypus-Opt==1.0.4
portalocker==2.10.1
posthog==3.7.0
POT==0.9.5
prettytable==3.12.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py4j==0.10.9.7
pyaml==24.9.0
pyaml-env==1.2.1
pyarrow==15.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycrypto==2.6.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.10.0
pylance==0.17.0
pynisher==1.0.10
pynndescent==0.5.13
pyparsing==3.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyrfr==0.9.0
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.12.0
seaborn==0.13.2
shap==0.46.0
shellingham==1.5.4
six==1.16.0
slicer==0.0.8
smac==2.2.0
smart-open==7.0.5
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.41.2
statsmodels==0.14.4
sympy==1.13.1
tblib==3.0.0
tenacity==8.5.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.3
tomli==2.1.0
toolz==1.0.0
torch==2.5.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
triton==3.1.0
typeguard==4.4.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
wrapt==1.16.0
yarl==1.17.1
zict==3.0.0
zipp==3.21.0
